Important changes for parsing ext4 + minor fixes for linux scenarios (#22)
1. The repeatable, but not consistent (across SWU files) failure to verify extraction of diffs at end of DiffGenTool.
This was due to an improper handling of content from ext4 file system in our code.
2. Incorrect package architectures for debian packages.
This simply wasn't implemented properly in our CMake files - we have now implemented checks using the compiler to get the target architecture.
3. recompress works properly in linux and has better error messages for failures
The issue here was differing requirements between linux and windows for popen() calls.
4. Changed name of the DiffGenTool archives to include version.
This is a nice to have and lets us move to have all the content in one output folder.

* Update source for 2.0.8

* Packages for 2.0.8